% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intchron.R
\name{intchron}
\alias{intchron}
\alias{intchron_hosts}
\title{Get data from IntChron}
\usage{
intchron(hosts, countries = NA, sites = NA, tabulate = TRUE)

intchron_hosts()
}
\arguments{
\item{hosts}{Databases to be retrieved. See details.}

\item{countries}{Vector of countries to be retrieved. See details.}

\item{sites}{Vector of sites to be retrieved. See details.}

\item{tabulate}{If \code{TRUE} (the default), the data retrieved will be combined
into a data frame. Set \code{FALSE} to get the raw data from IntChron.}
}
\value{
\code{intchron()}: a \code{tibble}, or if \code{tabulate = FALSE}, a list, of IntChron responses.

\code{intchron_hosts}: a \code{tibble} of available databases with their 'host' code.
}
\description{
IntChron is an indexing service for chronological data from multiple sources.
This function queries IntChron to retrieve data from the specified database
('host'), countries and sites.
}
\details{
At least \code{hosts} must be specified. Use \code{intchron_hosts()} to get a list of
currently available hosts, or "all" to retrieve data from any database.

See \code{vignette("rintchron")} for further details.
}
\examples{
# List available databases
intchron_hosts()

# Get all data from Jordan
intchron("all", countries = "Jordan")
}
