% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{intchron_crawl}
\alias{intchron_crawl}
\title{Recursively retrieve IntChron records}
\usage{
intchron_crawl(url, ignore = NA)
}
\arguments{
\item{url}{Address of the record to start crawling from.}

\item{ignore}{Vector of names of branches that should be ignored.}
}
\value{
A list of responses from IntChron.
}
\description{
Given an index record as a start point, this function recursively crawls the
tree of linked records and returns all the data records it finds.

The ignore parameter can be used to refine the search.
}
\details{
See \code{vignette("intchron-api")} for a detailed explanation of the structure of
the IntChron record tree and the crawling algorithm used.
}
\seealso{
Other functions for interacting with the IntChron API: 
\code{\link{intchron_extract}()},
\code{\link{intchron_request}()},
\code{\link{intchron_tabulate}()},
\code{\link{intchron_url}()}
}
\concept{functions for interacting with the IntChron API}
