% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{intchron_extract}
\alias{intchron_extract}
\title{Extract an element from an IntChron record}
\usage{
intchron_extract(x, what)
}
\arguments{
\item{x}{List of IntChron records from \code{\link[=intchron_request]{intchron_request()}}.}

\item{what}{Element to extract: \code{"data_url"}, \code{"records"}, \code{"file_data"}, or
\code{"series_file_data"}}
}
\value{
\code{what} from \code{x}.
}
\description{
Requests to IntChron return JSON responses that are mapped to a complex
nested list by \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}. This convenience function provides
access to specific named elements of those objects.
}
\examples{
khiv <- intchron_request(intchron_url("record", c("oxa", "nrcf"), "Jordan", "Kharaneh IV"))
intchron_extract(khiv, "data_url")
}
\references{
\itemize{
\item IntChron Data Schema \url{http://intchron.org/schema}
}
}
\seealso{
Other functions for interacting with the IntChron API: 
\code{\link{intchron_crawl}()},
\code{\link{intchron_request}()},
\code{\link{intchron_tabulate}()},
\code{\link{intchron_url}()}
}
\concept{functions for interacting with the IntChron API}
