% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intchron.R
\name{intchron_hosts}
\alias{intchron_hosts}
\alias{intchron_countries}
\title{List available databases and countries on IntChron}
\usage{
intchron_hosts()

intchron_countries(hosts = NA)
}
\arguments{
\item{hosts}{Vector of databases to query (for list of available countries).
Leave \code{NA} to list countries from all hosts.}
}
\value{
A \code{tibble} of available hosts or countries.
}
\description{
These functions query IntChron to list the available databases ('hosts') or
countries (optionally filtering by host).
}
\examples{
# List available hosts
intchron_hosts()

# List available countries
intchron_countries()

# List available countries for specific hosts
intchron_countries(c("intimate", "egyptdb"))
}
\seealso{
Other functions for querying IntChron: 
\code{\link{intchron}()}
}
\concept{functions for querying IntChron}
