% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intchron_api.R
\name{intchron_request}
\alias{intchron_request}
\title{Request resources from IntChron}
\usage{
intchron_request(url)
}
\arguments{
\item{url}{Vector of addresses to IntChron resources. See \code{\link[=intchron_url]{intchron_url()}}.}
}
\value{
A list of responses.
}
\description{
Gets the requested resources from IntChron and parses them with
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.
}
\examples{
intchron_request(intchron_url("record", "oxa", "Jordan", "Dhuweila"))
}
