% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intchron_api.R
\name{intchron_url}
\alias{intchron_url}
\title{Construct the address of IntChron resources}
\usage{
intchron_url(..., intchron_root = "https://intchron.org")
}
\arguments{
\item{...}{Character vectors describing the path to the resource or resources.}

\item{intchron_root}{Root address of IntChron.}
}
\value{
Character vector of URLs.
}
\description{
Combines a series of vectors describing the location of an IntChron resource
or resources into a properly encoded URL.
}
\examples{
# One resource
intchron_url("record", "oxa", "Jordan", "Dhuweila")

# Multiple resources
intchron_url("record", c("oxa", "nrcf"), "Jordan", "Kharaneh IV")
}
