% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_intchron_csv}
\alias{read_intchron_csv}
\title{Read a CSV file from IntChron}
\usage{
read_intchron_csv(file)
}
\arguments{
\item{file}{CSV records exported from IntChron; either a path to a downloaded
file, a URL, or literal data. Use \code{\link[readr:clipboard]{readr::clipboard()}} to read from the
system clipboard.}
}
\value{
A \code{tibble} containing the data from the record. Associated metadata is
discarded.
}
\description{
Reads records in the CSV format exported by by IntChron. These are regular
CSV files with a few elements of non-standard formatting that mean they can't
be directly parsed by e.g. \code{\link[=read.csv]{read.csv()}} or \code{\link[readr:read_delim]{readr::read_csv()}} (see details).

It is usually more robust to retrieve data from IntChron in JSON format using
\code{\link[=intchron]{intchron()}} or \code{\link[=intchron_request]{intchron_request()}}.
}
\details{
CSV files exported from IntChron have the following non-standard formatting:
\itemize{
\item Comment lines are denoted with '#' and contain metadata before and after
the table of data itself.
\item The comment line immediately above the data contains the column headings
\item A variable number of empty columns occur at the beginning of rows
\item A trailing comma occurs at the end of every row except the header
\item Missing values may be coded as: "", "-"
}

Beyond this, some data tables are malformed (e.g. they contain unmatched
quotes) and cannot be parsed without an error.
}
\concept{read and write functions}
