% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rintchron-package.R
\docType{package}
\name{rintchron-package}
\alias{rintchron}
\alias{rintchron-package}
\title{rintchron: Interface to IntChron}
\description{
An interface to 'IntChron' (\url{https://intchron.org)}), an indexing service and exchange format for chronological data such as radiocarbon dates (Bronk Ramsey et al. 2019) \doi{10.1017/RDC.2019.21}. Includes a user friendly interface for querying databases indexed by 'IntChron', as well as low level functions for interacting directly with the 'IntChron' API and other resources using its schema.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rintchron.joeroe.io}
  \item \url{https://github.com/joeroe/rintchron}
  \item Report bugs at \url{https://github.com/joeroe/rintchron/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

}
\keyword{internal}
