% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intchron.R
\name{intchron}
\alias{intchron}
\title{Query IntChron}
\usage{
intchron(hosts, countries = NA, sites = NA, tabulate = TRUE)
}
\arguments{
\item{hosts}{Vector of databases to be retrieved, or "all".}

\item{countries}{(Optional) Vector of countries to be retrieved.}

\item{sites}{(Optional) Vector of sites to be retrieved.}

\item{tabulate}{If \code{TRUE} (the default), the data retrieved will be combined
into a data frame. Set \code{FALSE} to get the raw data from IntChron.}
}
\value{
A \code{tibble}, or if \code{tabulate = FALSE}, a list, of IntChron responses.
}
\description{
IntChron is an indexing service for chronological data from multiple sources.
This function gets records from databases ('hosts') indexed by IntChron,
optionally filtering by country and/or site.
}
\details{
At least \code{hosts} must be specified. Use \code{intchron_hosts()} to get a list of
currently available hosts, or "all" to retrieve data from any database.

See \code{vignette("rintchron")} for further details.
}
\examples{
# Get data from Jordan from the ORAU and NERC databases
intchron(c("oxa", "nrcf"), countries = "Jordan")
}
\seealso{
Other functions for querying IntChron: 
\code{\link{intchron_hosts}()}
}
\concept{functions for querying IntChron}
