% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intchron.R
\name{intchron_get_labcode}
\alias{intchron_get_labcode}
\alias{intchron_get_doi}
\title{Retrieve individual records from IntChron}
\usage{
intchron_get_labcode(lab_code, tabulate = TRUE)

intchron_get_doi(doi, tabulate = TRUE)
}
\arguments{
\item{lab_code}{For radiocarbon dates, a vector of lab codes to be retrieved.}

\item{tabulate}{If \code{TRUE} (the default), the data retrieved will be combined
into a data frame. Set \code{FALSE} to get the raw responses from IntChron.}

\item{doi}{For publications, a vector of DOIs to be retrieved.}
}
\value{
A \code{tibble}, or if \code{tabulate = FALSE}, a list, of IntChron responses.
}
\description{
These functions retrieve individual records from IntChron by their primary
keys. Currently the two types of primary key supported by IntChron are
lab codes (for radiocarbon dates) and DOIs (for publications).
}
\examples{
intchron_get_labcode("OxA-18955")
}
\seealso{
Other functions for querying IntChron: 
\code{\link{intchron}()},
\code{\link{intchron_hosts}()}
}
\concept{functions for querying IntChron}
