% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{intchron_tabulate}
\alias{intchron_tabulate}
\title{Extract a data frame from IntChron}
\usage{
intchron_tabulate(x)
}
\arguments{
\item{x}{A list of IntChron responses.}
}
\value{
A data frame.
}
\description{
Takes a list of responses from an IntChron request and extracts the data
series and associated metadata as a data frame.
}
\examples{
khiv <- intchron_request(intchron_url("record", c("oxa", "nrcf"), "Jordan", "Kharaneh IV"))
intchron_tabulate(khiv)
}
\seealso{
Other functions for interacting with the IntChron API: 
\code{\link{intchron_crawl}()},
\code{\link{intchron_extract}()},
\code{\link{intchron_request}()},
\code{\link{intchron_url}()}
}
\concept{functions for interacting with the IntChron API}
